Rename and refactor packages
