+++ src/main/java/app/Application.java
@@ -1,16 +1,15 @@
-package app.deploy;
+package app;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.stereotype.Component;
 
 
 @Component
-@Configuration
-@ComponentScan(basePackages = "app")
-@EnableAutoConfiguration
+@SpringBootApplication
 public class Application {
 //    @Autowired
 //    private DataSource dataSource;
